+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Fixed
+- Proper error instead of panic, if specifying an input header file that doesn't exist.
+
+### Changed
+- If an output directory is specfied, it is recursively created, if it does not exist.
+  This can be avoided with --no-create-output-dir.
+- Output to stdout is now a complete header, similar to the contents of an output file
+  specified with --output-file. Earlier the output didn't contain comment, `#pragma once`
+  or `#include`s.
+
+
 ## [0.2.1] - 2025-08-29
 
 ### Fixed